Move auto-wrap code to a different file to solve a circular dependency.  Change uses of it to use the canonical path: lale.wrap_imported_operators
