+++ lale/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .helpers import wrap_imported_operators
+from .operator_wrapper import wrap_imported_operators
